Don't have docker containers rely on debug.log inode
Last night the bmon nodes stopped syncing because a debug.log logrotate
call happened, which changed the inode of the debug.log file. Because
the container volumes were to that specific file, the log watcher did
not detect the new file after rotation.

Fix this by mounting in the whole datadir.