Implement streaming MempoolAccept analytics
Including propagation delay across all hosts and within cohorts. The
metrics here are probably incomplete, and we're not persisting anything
to the database, but as-written this will at least give us something to
present on Grafana to establish an intuition for what data *would* be
interesting to persist and monitor.